+++ tests/conftest.py
@@ -1,0 +1,13 @@
+import pytest
+
+from legistar import base
+
+@pytest.fixture(scope="module")
+def scraper():
+    scraper = base.LegistarAPIScraper(None, None)
+    scraper.BASE_URL = 'http://webapi.legistar.com/v1/chicago'
+    scraper.retry_attempts = 0
+    scraper.requests_per_minute = 0
+    return scraper
+
+
